Fix warning about conversion. We might have to do it vice versa, though, comparing the float values and passing a converted number down to getMesh.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@4024 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CTriangleSelector.cpp
@@ -63,7 +63,7 @@ CTriangleSelector::CTriangleSelector(IAnimatedMeshSceneNode* node)
 	if (!animatedMesh)
 		return;
 
-	LastMeshFrame = AnimatedNode->getFrameNr();
+	LastMeshFrame = (u32)AnimatedNode->getFrameNr();
 	IMesh* mesh = animatedMesh->getMesh(LastMeshFrame);
 
 	if (mesh)
@@ -137,7 +137,7 @@ void CTriangleSelector::update(void) const
 	if (!AnimatedNode)
 		return; //< harmless no-op
 
-	const u32 currentFrame = AnimatedNode->getFrameNr();
+	const u32 currentFrame = (u32)AnimatedNode->getFrameNr();
 	if (currentFrame == LastMeshFrame)
 		return; //< Nothing to do
